management_service/requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
SQLAlchemy==1.4.47
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==0.21.1
alembic==1.10.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart